+++ CHANGELOG.md
@@ -15,6 +15,12 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 ## [Unreleased](https://github.com/hynek/svcs/compare/23.7.0...HEAD)
 
+### Added
+
+- It's now possible to request multiple services at once by passing multiple types to `Container.get()` and `Container.aget()`.
+  [#15](https://github.com/hynek/svcs/pull/15)
+
+
 
 ## [23.7.0](https://github.com/hynek/svcs/compare/23.6.0...23.7.0) - 2023-08-02